+++ eltakobus/util.py
@@ -1,10 +1,16 @@
+from enum import Enum
+
 def b2a(rawdata, separator=' '):
     # like binascii.b2a_hex, but directly to unicode for printing, and with nice spacing
     return separator.join("%02x"%b for b in rawdata)
 
 def b2s(rawdata, separator='-'):
     # like binascii.b2a_hex, but directly to unicode for printing, and with nice spacing
-    return separator.join("%02x"%b for b in rawdata).upper()
+    _rawdata = rawdata
+    if isinstance(_rawdata, AddressExpression):
+        _rawdata = _rawdata[0]
+    
+    return b2a(_rawdata, separator).upper()
 
 def combine_hex(data):
     ''' Combine list of integer values to one big integer '''
@@ -12,6 +18,9 @@ def combine_hex(data):
     for i, value in enumerate(reversed(data)):
         output |= (value << i * 8)
     return output
+
+def adr_plus_adr(a1:bytes, a2:bytes) -> bytes:
+    return (int.from_bytes(a1, 'big') + int.from_bytes(a2, 'big')).to_bytes(4, byteorder='big')
 
 class AddressExpression(tuple):
     """An object around the 4-long byte strings passed around for addressing
@@ -25,7 +34,7 @@ class AddressExpression(tuple):
         return "<%s %s>" % (type(self).__name__, self)
 
     def __str__(self):
-        return b2a(self[0]).replace(' ', '-') + (" %s" % self[1] if self[1] is not None else "")
+        return b2s(self[0]) + (" %s" % self[1] if self[1] is not None else "")
 
     @classmethod
     def parse(cls, s):
@@ -42,3 +51,60 @@ class AddressExpression(tuple):
         if self[1] is not None:
             raise ValueError("Address has disciminator %s, None expected" % self[1])
         return self[0]
+    
+    def is_local_address(self) -> bool:
+        return self[0][0:2] == b'\x00\x00'
+    
+    def add(self, address):
+        return AddressExpression((
+            (int.from_bytes(self[0], 'big') + int.from_bytes(address[0], 'big')).to_bytes(4, byteorder='big'),
+            self[1]
+        ))
+        
+
+class DefaultEnum(Enum):
+
+    # DEFAULT = (0, , 0'Unknown')
+
+    def __new__(cls, value, code:int=0, description:str=None):
+        obj = object.__new__(cls)
+        obj._value_ = value
+        obj._code = code
+        obj._description = description
+        return obj
+    
+    @classmethod
+    def find_by_code(cls, code):
+        for e in cls:
+            if e.code == code:
+                return e
+        return None
+    
+    @classmethod
+    def find_by_description(cls, description):
+        for e in cls:
+            if e.description == description:
+                return e
+        return None
+
+    @property
+    def value(self) -> int:
+        return self._value_
+
+    @property
+    def code(self) -> str:
+        return self._code
+
+    @property
+    def description(self) -> str:
+        return self._description
+        
+    def __repr__(self) -> str:
+        v_repr = self.__class__._value_repr_ or repr
+        repr = "<%s.%s: %s " % (self.__class__.__name__, self._name_, v_repr(self._value_))
+        if self.code:
+            repr += '"%S"' % (self.code)
+        if self.description:
+            repr += '"%S"' % (self.description)
+        repr += '>'
+        return repr